Task #31: Streamlining Ingestion Logic to account for Democracy Now
