+++ src/bootstrap.service.ts
@@ -1,13 +1,14 @@
 import { Injectable, OnApplicationBootstrap } from "@nestjs/common";
-import { SeasonGenService } from "./modules/seasons/seasonGenerator/seasonGen.service";
 import { IngestionTask } from "./modules/ingestion/ingestion.task";
 import { SeasonGenTask } from "./modules/seasons/seasonGenerator/seasonGen.task";
+import { NotificationsTask } from "./modules/notifications/notifications.task";
 
 @Injectable()
 export class BootstrapService implements OnApplicationBootstrap {
   constructor(
       private readonly ingestionTask: IngestionTask,
-      private readonly seasonGenTask: SeasonGenTask
+      private readonly seasonGenTask: SeasonGenTask,
+      private readonly notificationsTask: NotificationsTask
     ) {}
 
   async onApplicationBootstrap() {
@@ -17,5 +18,7 @@ export class BootstrapService implements OnApplicationBootstrap {
     await this.seasonGenTask.seasonGenerationTask();
     // 2. Sync shows from KDVS-API/SPINITRON then schedule notifications 
     await this.ingestionTask.runScheduleSync();
+    // 3. Schedule notifications for all shows until EOD
+    await this.notificationsTask.scheduleNotifications();
   }
 }
