+++ src/modules/ingestion/ingestion.module.ts
@@ -1,7 +1,6 @@
 import { Module } from '@nestjs/common';
 import { IngestionTask } from './ingestion.task';
 import { TimeslotsModule } from '../timeslots/timeslots.module';
-import { NotificationModule } from '../notifications/notifications.module';
 import { IngestionService } from './ingestion.service';
 import { PersonasModule } from '../personas/personas.module';
 import { SeasonsModule } from '../seasons/seasons.module';
@@ -13,7 +12,6 @@ import { ShowsModule } from '../shows/shows.module';
     SeasonsModule,
     PersonasModule,
     TimeslotsModule,
-    NotificationModule,
   ],
   providers: [
     IngestionService,
