+++ src/modules/ingestion/ingestion.service.ts
@@ -1,7 +1,5 @@
 import { Injectable, Logger } from "@nestjs/common";
 import { Show } from "../../shared/entities/show.entity";
-import { Season } from "../../shared/entities/season.entity";
-import { mergeZShowIntoShowMap } from "./kdvs-api/kdvs-api.helpers";
 import { extractZShowPersonaIds } from "./spinitron/spinitron.helper";
 import { ShowTimeslot } from "~/shared/entities/show-timeslot.entity";
 import { fetchZShowsForSeason } from "./kdvs-api/kdvs-api.client";
@@ -9,11 +7,10 @@ import { ShowsService } from "../shows/shows.service";
 import { fetchPersonasFromSpinitronIds } from "./spinitron/spinitron.client";
 import { PersonasService } from "../personas/personas.service";
 import { TimeslotsService } from "../timeslots/timeslots.service";
-import { appendZShowTimeslotByShowId, getLocalWeekday } from "../timeslots/timeslots.helper";
-import { zScheduleItem } from "./kdvs-api/kdvs-api.schema";
 import { DayOfWeek } from "~/shared/types/dotw.enum";
 import { TimeslotKey } from "../timeslots/timeslots.types";
 import { SeasonsService } from "../seasons/seasons.service";
+import { appendZShowTimeslotByShowId, mergeZShowIntoShowMap, recordZShowDOTW } from "./ingestion.helper";
 
 @Injectable()
 export class IngestionService {
@@ -25,22 +22,6 @@ export class IngestionService {
   ) {}
 
   private readonly logger = new Logger(IngestionService.name);
-
-  recordZShowDOTW(
-    showRecords: Map<string, Map<number, number>>, 
-    item: zScheduleItem
-  ) {
-    const showId = item.show_id ? String(item.show_id) : String(item.id);
-    const dotw = getLocalWeekday(item.start);
-    let showRecord = showRecords.get(showId);
-
-    if (!showRecord) {
-      showRecord = new Map<number, number>();
-      showRecords.set(showId, showRecord);
-    }
-
-    showRecord.set(dotw, (showRecord.get(dotw) ?? 0) + 1);
-  }
 
   async updateDB() {
     // Get the current Season if any
@@ -62,7 +43,7 @@ export class IngestionService {
     for (const zShow of zShows) {
       mergeZShowIntoShowMap(uniqueShowsById, zShow)
       extractZShowPersonaIds(uniquePersonaIds, zShow)
-      this.recordZShowDOTW(showWeekdayFrequency, zShow)
+      recordZShowDOTW(showWeekdayFrequency, zShow)
     }
 
     for (const zShow of zShows) {
