+++ src/modules/notifications/notifications.service.ts
@@ -119,4 +119,5 @@ export class NotificationService {
       deviceTokens,
     );
   }
+
 }
